refactor(device): extract channel/DIO/PWM/analog-out control from DaqifiStreamingDevice (part of #344)
The next extraction named in #344's status comment. `ChannelControlOperations`
(internal, `Device/Internal/`) now owns channel enable/disable and the ADC/DIO
enable-mask derivation, DIO direction and level, PWM enable/duty/frequency, and
the analog output. `DaqifiStreamingDevice` keeps the public surface and forwards.

Two members were added to the `IDeviceOperationHost` seam for this: the channels
snapshot and the channels lock. That was the reason #422 stopped short of this
block — mutating `IChannel.IsEnabled` and deriving the outbound mask must stay in
one critical section (#409), so the collaborator has to reach the device's lock
rather than take its own. It goes through the device's own members, so a
subclass's `Send` override still intercepts every command.

No public API change: the three PWM constants and `PwmFrequencyHz` stay on the
device with their docs, and every moved method keeps its signature, validation
order, and exception paramName.

Also removes `SendAdcEnableMask`/`SendDioEnableState`, two private methods left
uncalled since #411 replaced their call sites with the under-lock computation.

DaqifiStreamingDevice: 2,189 -> 1,756 lines.
